agents/agent_a/requirements.txt
pika>=1.2.0
redis
pytest
google-generativeai>=0.7.0
docker